+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Tests for "Walk on a die".
 - Tests for "Dolbear's law".
+- Tests for "ISBN check digit".
 
 ### Changed
 - Using "__dirname" for input / output paths.